:truck: [#52] Move and rename values table related files
The filenames have been updated for consistency with other module
names and the stories are now grouped under a folder 'values'
since there will be more functionality added to values.
